49 Find Commodity

+++ trade-module/src/main/java/io/edpn/backend/trade/infrastructure/persistence/entity/CommodityMarketInfoEntity.java
@@ -24,8 +24,8 @@ public class CommodityMarketInfoEntity {
     private Double minMeanPrice;
     private Double maxMeanPrice;
     private Double averageMeanPrice;
-    private Integer totalStock;
-    private Integer totalDemand;
+    private Long totalStock;
+    private Long totalDemand;
     private Integer totalStations;
     private Integer stationsWithBuyPrice;
     private Integer stationsWithSellPrice;
